refactor: consolidate api helpers
