+++ src/app/game-topup/page.tsx
@@ -2,7 +2,7 @@
 import GameTopUpPage from '@/components/topup/GameTopUpPage';
 import PageTransition from '@/components/animations/PageTransition';
 import { Metadata } from 'next';
-import { fetchJson } from '@/lib/fetchJson';
+import { fetchJson } from '@/lib/api/fetchJson';
 
 const app_name = process.env.NEXT_PUBLIC_APP_NAME;
 const app_url = process.env.NEXT_PUBLIC_BASE_URL;
