+++ src/lib/api/fetchJson.ts
@@ -1,8 +1,8 @@
-export async function fetchJson<T = any>(url: string, options?: RequestInit): Promise<T | null> {
+export async function fetchJson<T = unknown>(url: string, options?: RequestInit): Promise<T | null> {
   try {
     const res = await fetch(url, options);
     if (!res.ok) throw new Error('Failed to fetch');
-    return await res.json();
+    return (await res.json()) as T;
   } catch (e) {
     console.error(e);
     return null;
